Added interface function for creating projects
